pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "sightops"
version = "0.1.0"
description = "全屏视觉 AI 调研 Agent — 像真人一样操作电脑"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.30.0",
    "notion-client>=2.2.1",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.1",
    "typer>=0.12.3",
    "pyyaml>=6.0.1",
    "aiohttp>=3.9.5",
    "aiofiles>=23.2.1",
    "Pillow>=10.3.0",
    "loguru>=0.7.2",
    "httpx>=0.27.0",
    "rich>=13.7.1",
    "tenacity>=8.3.0",
    "python-slugify>=8.0.4",
    "humanize>=4.9.0",
    "pyautogui>=0.9.54",
]

[project.scripts]
sightops = "app.cli.app:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py311"
